Discretization Refactor with updated discretization and working process.

+++ Implementation/TimeInterval.py
@@ -1,7 +1,10 @@
+from ctypes import c_uint
+
+
 class TimeInterval:
     def __init__(self, start_point, end_point):
-        self.start_point = start_point
-        self.end_point = end_point
+        self.start_point = c_uint(start_point)
+        self.end_point = c_uint(end_point)
 
     def __lt__(self, other):
         return self.start_point < other.start_point or \
